fixed issue with search function
Search was one letter behind. The performSearch function was not
accepting any arguments. It was using the old this.props.value instead
of nextProps.query.

The performSearch from componentDidMount() now is passing
this.props.query

Then query was added to performSearch as an argument.

query replaced this.props.query  in the axios  call